requirements.txt
customtkinter
Pillow
matplotlib
bitarray